% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the sextant application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart, options, enableBookmarking, uiPattern}{Arguments passed to
[shiny::shinyApp()].}

\item{...}{arguments passed to `golem_opts`.}
}
\description{
Run the sextant application
}
