% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sextant.R
\name{sextant}
\alias{sextant}
\title{Draw a sextant}
\usage{
sextant(style = 1)
}
\arguments{
\item{style}{an integer controlling the style.  If this is 1,
which is the default, then medium-sized ticks are drawn every
5 degrees.  If it is 2, then they are drawn for even-numbered degrees.
The first style is similar to metric rulers, and the second
to Imperial rulers.}
}
\description{
Two styles are available, as selected by the \code{style} parameter.
}
\examples{
# Style 1
\dontrun{
pdf("sextant1.pdf", width=11, height=8.5, paper="USr")
sextant(1)
dev.off()
}
# Style 2
\dontrun{
pdf("sextant2.pdf", width=11, height=8.5, paper="USr")
sextant(2)
dev.off()
}
}
\author{
Dan Kelley
}
